-- 64,65

+++ src/main/scala/problems/P31.scala
@@ -11,11 +11,11 @@ import tools.Sandbox
 object P31 {
 
   // LAZY PRIMES GENERATION
-  def isPrime(x: Int): Boolean = (x > 1) && primes.takeWhile(_ <= math.sqrt(x)).forall(x % _ != 0)
+  def isPrime(x: Int): Boolean = (x > 1) && primes.takeWhile(_ <= math.sqrt(x.toDouble)).forall(x % _ != 0)
   val primes: LazyList[Int] = 2 #:: LazyList.from(3, 2).filter(isPrime)
 
   def isPrimeNaive(a: Int): Boolean =
-    (a > 1) && LazyList.from(2).take(math.sqrt(a).toInt + 1).forall(a % _ != 0)
+    (a > 1) && LazyList.from(2).take(math.sqrt(a.toDouble).toInt + 1).forall(a % _ != 0)
 }
 
 class P31 extends Sandbox {
